Custom click paramtypes moved to module
